[fix] - 모든 잘못된 내용 수정
