+++ app/question/page.tsx
@@ -1,5 +1,5 @@
 import TagSearchBar from "@/components/TagSearchBar";
-import QuestionCardList from "./components/QuestionCardList";
+import QuestionCardList from "../../features/question/components/QuestionCardList";
 import Paging from "@/components/common/Paging";
 
 const QuestionPage = () => {
